create platform-specific build tags for dropPrivileges() for metadata svc
