+++ lib/metadata/drop_privileges_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package metadata
+
+func dropPrivileges() (err error) {
+	return nil
+}
